Add status tracking for PrintingRequest and update order status
Introduced the PrintingRequestStatus enum and added a status field to PrintingRequest. Implemented updateStatusPrintingRequest in OrderBean and IOrderMgmt to update the status of individual printing requests and set the order status to FINISHED when all requests are in storage. Integrated this logic into ProductionBean to update printing request status upon storage completion.